+++ src/events/cps-mutation.ts
@@ -1,0 +1,9 @@
+type CpsMutation = CustomEvent<{ mutationList: MutationRecord[] }>;
+
+declare global {
+  interface GlobalEventHandlersEventMap {
+    'cps-mutation': CpsMutation;
+  }
+}
+
+export default CpsMutation;